Expose sql.DB's SetConnMaxLifetime in settings
When using a database server that closes inactive connections after some
duration, set `SQLConnMaxLifetime` to a value less than that duration.
Otherwise, the connection will be reused forever, and this could result in a
dropped connection.

For example, if you are using MySQL configured with `wait_timeout = 28800`, use
something like `SQLConnMaxLifetime=14400s`.

Refs #139