Add top of tag stack to error on repeating groups
Currently, debugging an incorrect repeating group template which is missing a tag is difficult because the required information (i.e. the tag which caused quickfix to think it had reached the end of the group) is not surfaced anywhere.
Adding the top tag of the stack to the error message makes debugging these issues much simpler.