Make Initiator`s inbound channel buffered (#22)
We have observed slow consumption of FIX messages in the Initiator.

Since the outbound channel is buffered, the event loop is sending too
many messages at once but reads incoming messages one by one. This
creates an inconsistency between incoming consumption messages and
outgoing production messages.

In this PR we resolve this issue by making the inbound channel buffered.
So the Initiator can now read and write in a truly asynchronous way.

This change is feature flagged. Can be used by setting
`InitiatorInChanCapacity` with an integer representing the capacity of
the inbound channel to the session settings config.